+++ pkg/migrations/rollback.go
@@ -1,0 +1,6 @@
+package migrations
+
+var (
+	// CapacityMarginFraction is the extra space required for upgrade rollback replica.
+	CapacityMarginFraction = 0.2
+)